requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
openai==1.69.0

# LangSmith integration
langsmith>=0.0.63

# Kubernetes client
kubernetes>=28.1.0

# Metrics and monitoring
requests>=2.31.0
requests-oauthlib>=1.3.1
oauthlib>=3.2.2

# API and Server
fastapi>=0.109.0
uvicorn>=0.25.0
pyyaml>=6.0.1
jinja2>=3.1.2

# Utilities
pydantic>=2.5.0
typeguard>=4.1.5
typing-extensions>=4.8.0
python-dotenv==1.0.1
rich>=13.7.0
python-dateutil==2.8.2
schema>=0.7.5
urllib3>=2.0.0
google-generativeai

# Visualization
matplotlib>=3.8.0
networkx>=3.2.1
pydot>=1.4.2

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-timeout>=2.2.0
pytest-mock>=3.12.0
